% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkinson.R
\name{atkinson}
\alias{atkinson}
\title{Atkinson Index based on Atkinson (1970)}
\usage{
atkinson(
  geo_large = "county",
  geo_small = "tract",
  year = 2020,
  subgroup,
  epsilon = 0.5,
  omit_NAs = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{geo_large}{Character string specifying the larger geographical unit of the data. The default is counties \code{geo_large = "county"}.}

\item{geo_small}{Character string specifying the smaller geographical unit of the data. The default is census tracts \code{geo_large = "tract"}.}

\item{year}{Numeric. The year to compute the estimate. The default is 2020, and the years 2009 onward are currently available.}

\item{subgroup}{Character string specifying the income or racial/ethnic subgroup(s) as the comparison population. See Details for available choices.}

\item{epsilon}{Numerical. Shape parameter that denotes the aversion to inequality. Value must be between 0 and 1.0 (the default is 0.5).}

\item{omit_NAs}{Logical. If FALSE, will compute index for a larger geographical unit only if all of its smaller geographical units have values. The default is TRUE.}

\item{quiet}{Logical. If TRUE, will display messages about potential missing census information. The default is FALSE.}

\item{...}{Arguments passed to \code{\link[tidycensus]{get_acs}} to select state, county, and other arguments for census characteristics}
}
\value{
An object of class 'list'. This is a named list with the following components:

\describe{
\item{\code{ai}}{An object of class 'tbl' for the GEOID, name, and AI at specified larger census geographies.}
\item{\code{ai_data}}{An object of class 'tbl' for the raw census values at specified smaller census geographies.}
\item{\code{missing}}{An object of class 'tbl' of the count and proportion of missingness for each census variable used to compute AI}
}
}
\description{
Compute the aspatial Atkinson Index of income or selected racial/ethnic subgroup(s) and U.S. geographies.
}
\details{
This function will compute the aspatial Atkinson Index (AI) of income or selected racial/ethnic subgroups and U.S. geographies for a specified geographical extent (e.g., the entire U.S. or a single state) based on Atkinson (1970) \doi{10.1016/0022-0531(70)90039-6}. This function provides the computation of AI for median household income and any of the U.S. Census Bureau race/ethnicity subgroups (including Hispanic and non-Hispanic individuals).

The function uses the \code{\link[tidycensus]{get_acs}} function to obtain U.S. Census Bureau 5-year American Community Survey characteristics used for the aspatial computation. The yearly estimates are available for 2009 onward when ACS-5 data are available but are available from other U.S. Census Bureau surveys. When \code{subgroup = "MedHHInc"}, the metric will be computed for median household income ("B19013_001"). The twenty racial/ethnic subgroups (U.S. Census Bureau definitions) are:
\itemize{
\item{B03002_002: }{not Hispanic or Latino "NHoL"}
\item{B03002_003: }{not Hispanic or Latino, white alone "NHoLW"}
\item{B03002_004: }{not Hispanic or Latino, Black or African American alone "NHoLB"}
\item{B03002_005: }{not Hispanic or Latino, American Indian and Alaska Native alone "NHoLAIAN"}
\item{B03002_006: }{not Hispanic or Latino, Asian alone "NHoLA"}
\item{B03002_007: }{not Hispanic or Latino, Native Hawaiian and Other Pacific Islander alone "NHoLNHOPI"}
\item{B03002_008: }{not Hispanic or Latino, Some other race alone "NHoLSOR"}
\item{B03002_009: }{not Hispanic or Latino, Two or more races "NHoLTOMR"}
\item{B03002_010: }{not Hispanic or Latino, Two races including Some other race "NHoLTRiSOR"}
\item{B03002_011: }{not Hispanic or Latino, Two races excluding Some other race, and three or more races "NHoLTReSOR"}
\item{B03002_012: }{Hispanic or Latino "HoL"}
\item{B03002_013: }{Hispanic or Latino, white alone "HoLW"}
\item{B03002_014: }{Hispanic or Latino, Black or African American alone "HoLB"}
\item{B03002_015: }{Hispanic or Latino, American Indian and Alaska Native alone "HoLAIAN"}
\item{B03002_016: }{Hispanic or Latino, Asian alone "HoLA"}
\item{B03002_017: }{Hispanic or Latino, Native Hawaiian and Other Pacific Islander alone "HoLNHOPI"}
\item{B03002_018: }{Hispanic or Latino, Some other race alone "HoLSOR"}
\item{B03002_019: }{Hispanic or Latino, Two or more races "HoLTOMR"}
\item{B03002_020: }{Hispanic or Latino, Two races including Some other race "HoLTRiSOR"}
\item{B03002_021: }{Hispanic or Latino, Two races excluding Some other race, and three or more races "HoLTReSOR"}
}

Use the internal \code{state} and \code{county} arguments within the \code{\link[tidycensus]{get_acs}} function to specify geographic extent of the data output.

AI is a measure of the evenness of residential inequality (e.g., racial/ethnic segregation) when comparing smaller geographical areas to larger ones within which the smaller geographical areas are located. The AI metric can range in value from 0 to 1 with smaller values indicating lower levels of inequality (e.g., less segregation).

The \code{epsilon} argument that determines how to weight the increments to inequality contributed by different proportions of the Lorenz curve. A user must explicitly decide how heavily to weight smaller geographical units at different points on the Lorenz curve (i.e., whether the index should take greater account of differences among areas of over- or under-representation). The \code{epsilon} argument must have values between 0 and 1.0. For \code{0 <= epsilon < 0.5} or less "inequality-averse," smaller geographical units with a subgroup proportion smaller than the subgroup proportion of the larger geographical unit contribute more to inequality ("over-representation"). For \code{0.5 < epsilon <= 1.0} or more "inequality-averse," smaller geographical units with a subgroup proportion larger than the subgroup proportion of the larger geographical unit contribute more to inequality ("under-representation"). If \code{epsilon = 0.5} (the default), units of over- and under-representation contribute equally to the index. See Section 2.3 of Saint-Jacques et al. (2020) \doi{10.48550/arXiv.2002.05819} for one method to select \code{epsilon}.

Larger geographies available include state \code{geo_large = "state"}, county \code{geo_large = "county"}, and census tract \code{geo_large = "tract"} levels. Smaller geographies available include, county \code{geo_small = "county"}, census tract \code{geo_small = "tract"}, and census block group \code{geo_small = "block group"} levels. If a larger geographical area is comprised of only one smaller geographical area (e.g., a U.S county contains only one census tract), then the AI value returned is NA.
}
\examples{
\dontrun{
# Wrapped in \dontrun{} because these examples require a Census API key.
  
  # Atkinson Index of non-Hispanic Black populations
  ## of census tracts within Georgia, U.S.A., counties (2020)
  atkinson(geo_large = "county", geo_small = "tract", state = "GA",
           year = 2020, subgroup = "NHoLB")
  
}

}
\seealso{
\code{\link[tidycensus]{get_acs}} for additional arguments for geographic extent selection (i.e., \code{state} and \code{county}).
}
